Fix type issue

+++ src/config/game.php
@@ -1,5 +1,5 @@
 <?php
 
 return [
-    'link_expiration_minutes' => env('GAME_LINK_EXPIRATION_MINUTES', 10080),
+    'link_expiration_minutes' => (int) env('GAME_LINK_EXPIRATION_MINUTES', 10080),
 ];
